+++ app/src/main/java/camp/nextstep/service/UserService.java
@@ -1,12 +1,42 @@
 package camp.nextstep.service;
 
+import camp.nextstep.dao.UserDao;
+import camp.nextstep.dao.UserHistoryDao;
 import camp.nextstep.domain.User;
+import camp.nextstep.domain.UserHistory;
+import com.interface21.beans.factory.annotation.Autowired;
+import com.interface21.context.stereotype.Service;
+import com.interface21.transaction.annotation.Transactional;
 
-import java.sql.SQLException;
+@Service
+public class UserService {
 
-public interface UserService {
+    private final UserDao userDao;
+    private final UserHistoryDao userHistoryDao;
 
-    User findById(final long id);
-    void save(final User user);
-    void changePassword(final long id, final String newPassword, final String createdBy) throws SQLException;
+    @Autowired
+    public UserService(final UserDao userDao, final UserHistoryDao userHistoryDao) {
+        this.userDao = userDao;
+        this.userHistoryDao = userHistoryDao;
+    }
+
+    public User findByAccount(final String account) {
+        return userDao.findByAccount(account);
+    }
+
+    public User findById(final long id) {
+        return userDao.findById(id);
+    }
+
+    public void save(final User user) {
+        userDao.insert(user);
+    }
+
+    @Transactional
+    public void changePassword(final long id, final String newPassword, final String createBy) {
+        final var user = findById(id);
+        user.changePassword(newPassword);
+        userDao.update(user);
+        userHistoryDao.log(new UserHistory(user, createBy));
+    }
 }
